validation_pipeline/requirements.txt
# Validation Pipeline Dependencies

# Google Gemini API
google-generativeai>=0.3.0

# PDF processing
PyPDF2>=3.0.0

# Environment variables
python-dotenv>=1.0.0

# JSON handling (usually included in Python, but listed for clarity)
# json - built-in

